frontend: Rename to priorityClasses to priorityClass
The name for this resource was pluralized as a mistake. The previous
module is still shipped for compatibility reasons.
